Auto Rigging VR5
This is my Version 5, where I solve a problem with the FK controllers, so when you create you can already have a parentConstraint. The function "def parentCTRL()" was my first test for this function.
The goal is to make the controllers for IK, make an option for user to choose the name for the joints and blending IK/FK.